backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.31
pandas-ta>=0.3.14b1
scipy>=1.11.0
pydantic>=2.0.0
